+++ CHANGES.md
@@ -3,6 +3,7 @@ Release notes and TODO items.
 
 ## 0.1.2 / 2016-July-??
 
+* Fix wrong driver classname in Derby network connections
 * Validate test query as either string or unspecified
 * Infer adapter name for implicit types